refactor(monitor): add class View
